--- a/lib/polylines.rb
+++ b/lib/polylines.rb
@@ -1,4 +1,3 @@
-require "polylines/base"
 require "polylines/encoder"
 require "polylines/decoder"