… lib/polylines/base.rb
… spec/polylines/encoder_spec.rb