Hi Joshua,

We love using your gem @wayzup ! But we encountered some performances issues. I decided to try a faster implementation of the polyline encoding / decoding algorithm while respecting your interfaces (& the ability to encode / decode a single coordinate). I was highly inspirated by the work of Jan Varwig (https://gist.github.com/janv/1361880) & Edward Samuel (https://github.com/edwardsamuel/google-maps-services-ruby).  what do you think of it ? 

A quick benchmark :

````
require 'benchmark'

TURNS = 50_000
POLY = '_p~iF~ps|U_ulLnnqC_mqNvxq`@'
POLY_ARR = [[38.5, -120.2], [40.7, -120.95], [43.252, -126.453]]

def benchmark(&_block)
  bench = Benchmark.measure { yield }
  puts "=> Time spent: #{bench.real.round(2)} s"
end

gem "polylines", git: 'git://github.com/joshuaclayton/polylines.git'
require "polylines"

puts "EXISTING - Polylines::Decoder.decode_polyline"
benchmark {
  TURNS.times do
    Polylines::Decoder.decode_polyline(POLY)
  end
}

puts "EXISTING - Polylines::Encoder.encode_points"
benchmark {
  TURNS.times do
    Polylines::Encoder.encode_points(POLY_ARR)
  end
}

Object.send(:remove_const, :Polylines)

require_relative "lib/polylines/decoder"
require_relative "lib/polylines/encoder"

puts "NEW - Polylines::Decoder.decode_polyline"
benchmark {
  TURNS.times do
    Polylines::Decoder.decode_polyline(POLY)
  end
}

puts "NEW - Polylines::Encoder.encode_points"
benchmark {
  TURNS.times do
    Polylines::Encoder.encode_points(POLY_ARR)
  end
}
````

Gives the following results on my machine (Macbook pro 13", 2015) : 

`````
EXISTING - Polylines::Decoder.decode_polyline
=> Time spent: 10.6 s
EXISTING - Polylines::Encoder.encode_points
=> Time spent: 2.96 s
NEW - Polylines::Decoder.decode_polyline
=> Time spent: 0.88 s
NEW - Polylines::Encoder.encode_points
=> Time spent: 0.4 s
`````