@joshuaclayton 2017-03-19
@ccyrille thanks for reaching out!

One of the reasons I wrote the gem in the fashion I did was to isolate the steps in a way that made sense to me; I recall running into other examples written in Ruby but maybe ran into an issue and couldn't figure out how to debug.

While I see the argument for performance, the current version is already sub-ms for small- to medium-sized datasets and other mechanisms (e.g. https://github.com/joshuaclayton/douglas_peucker) could be used to simplify lines prior to encoding to reduce overall number of points. Should larger datasets (e.g. 10,000 coordinate pairs) need to be encoded, I imagine that work could be moved to a background job so it's not happening in a web request.

I don't plan on accepting this as-is (given there are other source examples that are also performant), as I want to be able to maintain this over time. However, if there were some quick performance wins within the structure of the existing multi-step setup, I'd be interested in that.

Thanks again for your interest and the vote of confidence in the gem.