@ccyrille 2017-03-19
@joshuaclayton Thanks for taking the time to review this implementation !

I perfectly understand your point of view : you implementation is currently much more didactic that the one I offer. But I didn't find how to drastically improve the speed without breaking the structure. 

As we particularly need this @wayzup, I am going to release it as a separate implementation :)